+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 
 - Changed Documentation and Accessibility pages of the GUI to work as Flask templates
 - Cleaned up Flask files to not be all in one module
+- Changed the database functionality to store possible conversions as a graph instead of a table
 
 ## v0.1.6